+++ static/app/components/issues/suspect/constants.tsx
